% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday Function}
\usage{
birthday(x)
}
\arguments{
\item{x}{Random sample of size x of people}
}
\value{
The probability of two of more people in x having a common birthday
}
\description{
Birthday Function
}
\examples{
\dontrun{birthday(23,365)}
}
