% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM DATA}
\format{
A data frame with 16 rows and 2 columns:
\describe{
  \item{DISTANCE}{Distance from fire}
  \item{DAMAGE}{Damage score or severity}
}
}
\source{
Local dataset "FIREDAM.csv"
}
\usage{
fire
}
\description{
A subset of data showing fire distance and damage measurements.
}
\keyword{datasets}
