% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of iterations}

\item{x}{Sample data}

\item{fun}{Statistic of bootstrap}

\item{alpha}{Significance level}

\item{cx}{size of the ci text on histogram}

\item{...}{Extra plotting arguments}
}
\value{
A list
}
\description{
myboot2
}
\examples{
\dontrun{myboot2(iter=10000)}
}
