% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Sample size per iteration}

\item{iter}{Number of iterations}

\item{lambda}{Theoretical Poisson probability function}

\item{...}{Additional parameters}
}
\value{
Vector of sample means and greats histograms
}
\description{
mycltp
}
\examples{
\dontrun{mycltp(n=10,iter=10000)}
}
