% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{myhyper function}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{number of times repeated}

\item{N}{total population size}

\item{r}{number of successions in population}

\item{n}{sample size}
}
\value{
probability distribution
}
\description{
myhyper function
}
\examples{
\dontrun{myhyper(10, 5, 3)}
}
