% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal distribution}

\item{sigma}{Standard deviation of the normal distribution}

\item{a}{Value up to which to chade and calculate probability}
}
\value{
a list of the results containing mu, sigma, a and probability
}
\description{
myncurve
}
\examples{
\dontrun{myncurve(1, 2, 3)}
}
