% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats available on flight}

\item{gamma}{Acceptable probability of overbooking}

\item{p}{Probability that a passenger with a ticket shows}
}
\value{
A list containing nd, nc, N, p, and gamma. Also returns two plots
}
\description{
ntickets
}
\examples{
\dontrun{ntickets(N=400,gamma = 0.02, p = 0.95)}
}
